[wip] display each author update in a separate comments section
