+++ src/containers/Proposal/Detail/hooks.js
@@ -1,4 +1,4 @@
-import { useMemo, useState } from "react";
+import { useMemo, useState, useEffect, useCallback } from "react";
 import * as sel from "src/selectors";
 import * as act from "src/actions";
 import { useSelector, useAction } from "src/redux";
@@ -7,9 +7,13 @@ import {
   getProposalToken,
   getTokensForProposalsPagination
 } from "../helpers";
-import { getDetailsFile } from "./helpers";
+import { getDetailsFile, calculateAuthorUpdateTree } from "./helpers";
 import { shortRecordToken, parseRawProposal } from "src/helpers";
-import { PROPOSAL_STATE_VETTED } from "src/constants";
+import {
+  PROPOSAL_STATE_VETTED,
+  PROPOSAL_UPDATE_HINT,
+  PROPOSAL_MAIN_THREAD_KEY
+} from "src/constants";
 import useFetchMachine from "src/hooks/utils/useFetchMachine";
 import { useLoaderContext } from "src/containers/Loader";
 import isEmpty from "lodash/fp/isEmpty";
@@ -19,6 +23,8 @@ import isEqual from "lodash/fp/isEqual";
 import values from "lodash/fp/values";
 import pick from "lodash/pick";
 import concat from "lodash/fp/concat";
+import { useConfig } from "src/containers/Config";
+import { or } from "src/lib/fp";
 
 const getUnfetchedVoteSummaries = (proposal, voteSummaries, isSubmission) => {
   if (!proposal) return [];
@@ -187,5 +193,156 @@ export function useProposal(token, threadParentID) {
     loading: state.status === "idle" || state.status === "loading",
     threadParentID,
     isCurrentUserProposalAuthor
+  };
+}
+
+export function useComments(recordToken, proposalState) {
+  const { enableCommentVote, recordType, constants } = useConfig();
+
+  const errorSelector = or(
+    sel.apiProposalCommentsError,
+    sel.apiInvoiceCommentsError,
+    sel.apiDccCommentsError,
+    sel.apiLikeCommentsError,
+    sel.apiCommentsLikesError
+  );
+  const error = useSelector(errorSelector);
+  const commentsSelector = useMemo(
+    () => sel.makeGetRecordComments(recordToken),
+    [recordToken]
+  );
+  const commentsLikesSelector = useMemo(
+    () => sel.makeGetRecordCommentsLikes(recordToken),
+    [recordToken]
+  );
+  const commentsVotesSelector = useMemo(
+    () => sel.makeGetRecordCommentsVotes(recordToken),
+    [recordToken]
+  );
+  const lastVisitTimestampSelector = useMemo(
+    () => sel.makeGetLastAccessTime(recordToken),
+    [recordToken]
+  );
+  const comments = useSelector(commentsSelector);
+  const commentsLikes = useSelector(commentsLikesSelector);
+  const commentsVotes = useSelector(commentsVotesSelector);
+  const lastVisitTimestamp = useSelector(lastVisitTimestampSelector);
+  const loading = useSelector(sel.isApiRequestingComments);
+  const loadingLikes = useSelector(sel.isApiRequestingCommentsLikes);
+  const onSubmitComment = useAction(
+    recordType === constants.RECORD_TYPE_DCC
+      ? act.onSaveNewDccComment
+      : act.onSaveNewComment
+  );
+  const onFetchComments = useAction(
+    recordType === constants.RECORD_TYPE_PROPOSAL
+      ? act.onFetchProposalComments
+      : recordType === constants.RECORD_TYPE_DCC
+      ? act.onFetchDccComments
+      : act.onFetchInvoiceComments
+  );
+  const onFetchLikes = useAction(act.onFetchLikedComments);
+  const onCommentVoteAction = useAction(act.onCommentVote);
+  const onCensorComment = useAction(act.onCensorComment);
+
+  const { currentUser } = useLoaderContext();
+  const userid = currentUser && currentUser.userid;
+  const email = currentUser && currentUser.email;
+
+  const userLoggedIn = !!email;
+
+  // comments are not public on cms. User needs to be logged in
+  const isProposal = recordType === constants.RECORD_TYPE_PROPOSAL;
+  const needsToFetchComments = isProposal
+    ? !!recordToken && !comments
+    : !!recordToken && !comments && userLoggedIn;
+
+  const needsToFetchCommentsLikes =
+    !!recordToken &&
+    !commentsLikes &&
+    enableCommentVote &&
+    userLoggedIn &&
+    proposalState === PROPOSAL_STATE_VETTED;
+
+  useEffect(
+    function handleFetchOfComments() {
+      if (needsToFetchComments) {
+        onFetchComments(recordToken);
+      }
+    },
+    [onFetchComments, needsToFetchComments, recordToken]
+  );
+
+  useEffect(
+    function handleFetchOfLikes() {
+      if (needsToFetchCommentsLikes) {
+        onFetchLikes(recordToken, userid);
+      }
+    },
+    [onFetchLikes, needsToFetchCommentsLikes, recordToken, userid]
+  );
+
+  const onCommentVote = useCallback(
+    (commentID, action, token) => {
+      onCommentVoteAction(userid, token, commentID, action, proposalState);
+    },
+    [onCommentVoteAction, userid, proposalState]
+  );
+
+  const getCommentLikeOption = useCallback(
+    (commentID) => {
+      const option = commentsLikes && commentsLikes[commentID];
+      return option ? option : 0;
+    },
+    [commentsLikes]
+  );
+
+  const getCommentVotes = useCallback(
+    (commentID) => commentsVotes && commentsVotes[commentID],
+    [commentsVotes]
+  );
+
+  const authorUpdateIds = comments
+    ?.filter(({ extradatahint }) => extradatahint === PROPOSAL_UPDATE_HINT)
+    .map(({ commentid }) => commentid);
+
+  // Calculate comments tree for each author update to display each
+  // one of them in a separate comments section.
+  const commentsWithAuthorUpdatesMap = useMemo(() => {
+    const commentsMap = {};
+    let authorUpdateThreads = [];
+    if (authorUpdateIds) {
+      authorUpdateIds.forEach((updateId) => {
+        const authorUpdateTree = calculateAuthorUpdateTree(updateId, comments);
+        authorUpdateThreads = [...authorUpdateThreads, ...authorUpdateTree];
+        commentsMap[updateId] = comments.filter(({ commentid }) =>
+          authorUpdateTree.includes(commentid)
+        );
+      });
+      commentsMap[PROPOSAL_MAIN_THREAD_KEY] = comments.filter(
+        ({ commentid }) => !authorUpdateThreads.includes(commentid)
+      );
+    }
+    return commentsMap;
+  }, [authorUpdateIds, comments]);
+
+  return {
+    comments: authorUpdateIds ? commentsWithAuthorUpdatesMap : comments,
+    onCommentVote,
+    onCensorComment,
+    getCommentLikeOption,
+    enableCommentVote,
+    userLoggedIn,
+    userEmail: email,
+    recordType,
+    currentUser,
+    lastVisitTimestamp,
+    loading,
+    loadingLikes,
+    onSubmitComment,
+    error,
+    getCommentVotes,
+    authorUpdateIds,
+    latestAuthorUpdateId: authorUpdateIds && authorUpdateIds[0]
   };
 }
